backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
pydantic
langchain-community
